Fix all 29 mypy type errors
- Cast numpy/pandas return values to explicit types (float, bool, np.ndarray)
- Narrow Anthropic SDK response union type before accessing .text
- Add return type annotations to all FastAPI endpoints
- Add missing get_all_features() and health_check() to FeatureStore
- Use string colormap name instead of plt.cm.Blues
- Add types-requests to dev dependencies

+++ src/features/feature_engineering.py
@@ -22,7 +22,7 @@ def tx_frequency_per_day(txs: pd.DataFrame) -> float:
     span = (dates.max() - dates.min()).total_seconds() / 86_400
     if span == 0:
         return float(len(txs))
-    return len(txs) / span
+    return float(len(txs) / span)
 
 
 def activity_regularity(txs: pd.DataFrame) -> float:
